code climate refactor fix more issues
